pyproject.toml
[project]
name = "health-wellness-planner-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "google-api-python-client>=2.175.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "ics>=0.7.2",
    "openai-agents>=0.1.0",
    "pydantic-ai[logfire]>=0.3.6",
    "python-dotenv>=1.1.1",
]
